requirements.txt
alabaster==0.7.13
altair==4.2.0
attrs==22.1.0
Babel==2.11.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
decorator==5.1.1
diffusers==0.2.4
docutils==0.19
dulwich==0.20.50
entrypoints==0.4
filelock==3.8.0
ftfy==6.1.1
gitdb==4.0.9
GitPython==3.1.29
huggingface-hub==0.9.0
idna==3.4
imagesize==1.4.1
imaplib2==3.6
importlib-metadata==5.0.0
importlib-resources==5.10.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
livereload==2.6.3
MarkupSafe==2.1.1
numpy==1.19.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.5.5.64
openvino==2022.1.0
packaging==21.3
pandas==1.4.4
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
progressbar==2.5
protobuf==3.20.3
pyarrow==10.0.0
pydantic==1.10.4
pydeck==0.8.0b4
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rstcheck==6.1.1
rstcheck-core==1.0.3
sacremoses==0.0.53
schedule==1.2.0
scipy==1.9.0
semver==2.13.0
shellingham==1.5.0.post1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==6.1.3
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
streamlit==1.12.0
tokenizers==0.13.1
toml==0.10.2
toolz==0.12.0
torch==1.13.0
tornado==6.2
tqdm==4.64.0
transformers==4.16.2
tweepy==4.14.0
typer==0.7.0
types-docutils==0.19.1.3
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
zipp==3.10.0
